more bug fixes targeting NaN in loss; iq_accum was fp16, changed to fp32; MessagePass now is guarded against 0/0 in backwards pass; loss inputs fcast to float32

+++ ScatterNet/train/loss.py
@@ -76,9 +76,7 @@ class Loss(nn.Module):
         Returns:
             per-molecule per-q losses, shape (N, Q)
         """
-        # cast to fp32 before log1p: for large molecules I(q) >> fp16 max (~65504),
-        # causing both pred and target to overflow to +inf, then inf-inf = NaN.
-        residual = torch.log1p(output_head.float()) - torch.log1p(batch.iqval.float())
+        residual = torch.log1p(output_head) - torch.log1p(batch.iqval.float())
         return self._q_weights_ * residual ** 2
 
     @jaxtyped(typechecker=beartype)
@@ -121,15 +119,21 @@ class Loss(nn.Module):
         output_head: Float[torch.Tensor, "N Q"],
         f_mag_pred:  Float[torch.Tensor, "N M Q"],
         sigma_pred:  Float[torch.Tensor, "N M Q"],
-        batch: Batch, 
+        batch: Batch,
         lambda_6: float,
         lambda_7: float,
-        eps: float 
+        eps: float
     ) -> Float[torch.Tensor, ""]:
-        
+
+        # cast to float32 before any loss computation: fp16 can overflow or produce
+        # inf which then causes NaN in backward (inf*0, 0/0 patterns).
+        output_head = output_head.float()
+        f_mag_pred  = f_mag_pred.float()
+        sigma_pred  = sigma_pred.float()
+
         msle_loss  = self._kratky_MSLE(output_head, batch)
         ff_penalty = self._ff_penalty(f_mag_pred, batch, lambda_6)
         sg_penalty = self._sg_penalty(sigma_pred, lambda_7, batch, eps)
-        
+
         return (msle_loss + ff_penalty + sg_penalty).mean()
